Refactor utility functions into utility class
